Use JWKS URL instead of OIDC discovery in vault
Vault expects the same certificates are used for OIDC discovery and
JWKS URL, which is not our case, so vault fails to validate the host
behind JWKS endpoint.

This change sets directly the JWKS URL, and Vault will use the system
certificates to check the host (let's encrypt in our case).